requirements.txt
pillow
pathlib 